+++ src/tests/sqlsmith/src/lib.rs
@@ -12,14 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![feature(once_cell)]
 #![feature(let_chains)]
+#![feature(once_cell)]
 
 use std::vec;
 
 use itertools::Itertools;
 use rand::prelude::SliceRandom;
 use rand::Rng;
+use risingwave_common::row::Row2;
 use risingwave_common::types::DataTypeName;
 use risingwave_frontend::bind_data_type;
 use risingwave_sqlparser::ast::{
@@ -28,456 +29,14 @@ use risingwave_sqlparser::ast::{
 };
 use risingwave_sqlparser::parser::Parser;
 
-mod expr;
-pub use expr::print_function_table;
-use risingwave_expr::ExprError;
+use crate::sql_gen::SqlGenerator;
 
-mod relation;
 pub mod runner;
-mod scalar;
-mod time_window;
-mod utils;
-use crate::utils::create_table_with_joins_from_table;
+mod sql_gen;
+pub mod validation;
+pub use validation::is_permissible_error;
 
-#[derive(Clone, Debug)]
-pub struct Table {
-    pub name: String,
-    pub columns: Vec<Column>,
-}
-
-impl Table {
-    pub fn new(name: String, columns: Vec<Column>) -> Self {
-        Self { name, columns }
-    }
-
-    pub fn get_qualified_columns(&self) -> Vec<Column> {
-        self.columns
-            .iter()
-            .map(|c| Column {
-                name: format!("{}.{}", self.name, c.name),
-                data_type: c.data_type,
-            })
-            .collect()
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct Column {
-    name: String,
-    data_type: DataTypeName,
-}
-
-impl From<ColumnDef> for Column {
-    fn from(c: ColumnDef) -> Self {
-        Self {
-            name: c.name.real_value(),
-            data_type: bind_data_type(&c.data_type).unwrap().into(),
-        }
-    }
-}
-
-#[derive(Copy, Clone)]
-struct SqlGeneratorContext {
-    can_agg: bool, // This is used to disable agg expr totally,
-                   // Used in top level, where we want to test queries
-                   // without aggregates.
-    inside_agg: bool, // TODO: Maybe we can get rid of this???
-    inside_explicit_cast: bool,
-}
-
-impl SqlGeneratorContext {
-    pub fn new() -> Self {
-        SqlGeneratorContext {
-            can_agg: true,
-            inside_agg: false,
-            inside_explicit_cast: false,
-        }
-    }
-    pub fn new_with_can_agg(can_agg: bool) -> Self {
-        Self {
-            can_agg,
-            inside_agg: false,
-            inside_explicit_cast: false,
-        }
-    }
-    pub fn set_inside_agg(self) -> Self {
-        Self {
-            inside_agg: true,
-            ..self
-        }
-    }
-    pub fn set_inside_explicit_cast(self) -> Self {
-        Self {
-            inside_explicit_cast: true,
-            ..self
-        }
-    }
-    pub fn can_implicit_cast(self) -> bool {
-        !self.inside_explicit_cast
-    }
-    pub fn can_gen_agg(self) -> bool {
-        self.can_agg && !self.inside_agg
-    }
-}
-
-struct SqlGenerator<'a, R: Rng> {
-    tables: Vec<Table>,
-    rng: &'a mut R,
-
-    /// Relation ID used to generate table names and aliases
-    relation_id: u32,
-
-    /// is_distinct_allowed - Distinct and Orderby/Approx.. cannot be generated together among agg
-    ///                       having and
-    /// When this variable is true, it means distinct only
-    /// When this variable is false, it means orderby and approx only.
-    is_distinct_allowed: bool,
-
-    /// Relations bound in generated query.
-    /// We might not read from all tables.
-    bound_relations: Vec<Table>,
-
-    /// Columns bound in generated query.
-    /// May not contain all columns from Self::bound_relations.
-    /// e.g. GROUP BY clause will constrain bound_columns.
-    bound_columns: Vec<Column>,
-
-    /// SqlGenerator can be used in two execution modes:
-    /// 1. Generating Query Statements.
-    /// 2. Generating queries for CREATE MATERIALIZED VIEW.
-    ///    Under this mode certain restrictions and workarounds are applied
-    ///    for unsupported stream executors.
-    is_mview: bool,
-}
-
-/// Generators
-impl<'a, R: Rng> SqlGenerator<'a, R> {
-    fn new(rng: &'a mut R, tables: Vec<Table>) -> Self {
-        let is_distinct_allowed = rng.gen_bool(0.5);
-        SqlGenerator {
-            tables,
-            rng,
-            relation_id: 0,
-            is_distinct_allowed,
-            bound_relations: vec![],
-            bound_columns: vec![],
-            is_mview: false,
-        }
-    }
-
-    fn new_for_mview(rng: &'a mut R, tables: Vec<Table>) -> Self {
-        // distinct aggregate is not allowed for MV
-        SqlGenerator {
-            tables,
-            rng,
-            relation_id: 0,
-            is_distinct_allowed: false,
-            bound_relations: vec![],
-            bound_columns: vec![],
-            is_mview: true,
-        }
-    }
-
-    fn add_relations_to_context(&mut self, mut tables: Vec<Table>) {
-        for rel in &tables {
-            let mut bound_columns = rel.get_qualified_columns();
-            self.bound_columns.append(&mut bound_columns);
-        }
-        self.bound_relations.append(&mut tables);
-    }
-
-    fn gen_stmt(&mut self) -> Statement {
-        let (query, _) = self.gen_query();
-        Statement::Query(Box::new(query))
-    }
-
-    pub fn gen_mview(&mut self, name: &str) -> (Statement, Table) {
-        let (query, schema) = self.gen_query();
-        let query = Box::new(query);
-        let table = Table {
-            name: name.to_string(),
-            columns: schema,
-        };
-        let name = ObjectName(vec![Ident::new(name)]);
-        let mview = Statement::CreateView {
-            or_replace: false,
-            materialized: true,
-            name,
-            columns: vec![],
-            query,
-            with_options: vec![],
-        };
-        (mview, table)
-    }
-
-    /// Generates query expression and returns its
-    /// query schema as well.
-    fn gen_query(&mut self) -> (Query, Vec<Column>) {
-        if self.can_recurse() {
-            self.gen_complex_query()
-        } else {
-            self.gen_simple_query()
-        }
-    }
-
-    /// Generates a complex query which may recurse.
-    /// e.g. through `gen_with` or other generated parts of the query.
-    fn gen_complex_query(&mut self) -> (Query, Vec<Column>) {
-        let (with, with_tables) = self.gen_with();
-        let (query, schema) = self.gen_set_expr(with_tables);
-        (
-            Query {
-                with,
-                body: query,
-                order_by: self.gen_order_by(),
-                limit: self.gen_limit(),
-                offset: None,
-                fetch: None,
-            },
-            schema,
-        )
-    }
-
-    /// Generates a simple query which will not recurse.
-    fn gen_simple_query(&mut self) -> (Query, Vec<Column>) {
-        let with_tables = vec![];
-        let (query, schema) = self.gen_set_expr(with_tables);
-        (
-            Query {
-                with: None,
-                body: query,
-                order_by: vec![],
-                limit: None,
-                offset: None,
-                fetch: None,
-            },
-            schema,
-        )
-    }
-
-    /// Generates a query with local context.
-    /// Used by `WITH`, `Table Subquery` in Relation
-    fn gen_local_query(&mut self) -> (Query, Vec<Column>) {
-        let old_ctxt = self.new_local_context();
-        let t = self.gen_query();
-        self.restore_context(old_ctxt);
-        t
-    }
-
-    /// Generates a query with correlated context to ensure proper recursion.
-    /// Used by Exists `Subquery`
-    /// TODO: <https://github.com/risingwavelabs/risingwave/pull/4431#issuecomment-1327417328>
-    fn _gen_correlated_query(&mut self) -> (Query, Vec<Column>) {
-        let old_ctxt = self._clone_local_context();
-        let t = self.gen_query();
-        self.restore_context(old_ctxt);
-        t
-    }
-
-    fn gen_with(&mut self) -> (Option<With>, Vec<Table>) {
-        match self.flip_coin() {
-            true => (None, vec![]),
-            false => {
-                let (with, tables) = self.gen_with_inner();
-                (Some(with), tables)
-            }
-        }
-    }
-
-    fn gen_with_inner(&mut self) -> (With, Vec<Table>) {
-        let alias = self.gen_table_alias_with_prefix("with");
-        let (query, query_schema) = self.gen_local_query();
-        let from = None;
-        let cte = Cte {
-            alias: alias.clone(),
-            query,
-            from,
-        };
-
-        let with_tables = vec![Table {
-            name: alias.name.real_value(),
-            columns: query_schema,
-        }];
-        (
-            With {
-                recursive: false,
-                cte_tables: vec![cte],
-            },
-            with_tables,
-        )
-    }
-
-    fn gen_set_expr(&mut self, with_tables: Vec<Table>) -> (SetExpr, Vec<Column>) {
-        match self.rng.gen_range(0..=9) {
-            0..=9 => {
-                let (select, schema) = self.gen_select_stmt(with_tables);
-                (SetExpr::Select(Box::new(select)), schema)
-            }
-            _ => unreachable!(),
-        }
-    }
-
-    fn gen_order_by(&mut self) -> Vec<OrderByExpr> {
-        if self.bound_columns.is_empty() || !self.is_distinct_allowed {
-            return vec![];
-        }
-        let mut order_by = vec![];
-        while self.flip_coin() {
-            let column = self.bound_columns.choose(&mut self.rng).unwrap();
-            order_by.push(OrderByExpr {
-                expr: Expr::Identifier(Ident::new(&column.name)),
-                asc: Some(self.rng.gen_bool(0.5)),
-                nulls_first: None,
-            })
-        }
-        order_by
-    }
-
-    fn gen_limit(&mut self) -> Option<String> {
-        if !self.is_mview && self.rng.gen_bool(0.2) {
-            Some(self.rng.gen_range(0..=100).to_string())
-        } else {
-            None
-        }
-    }
-
-    fn gen_select_stmt(&mut self, with_tables: Vec<Table>) -> (Select, Vec<Column>) {
-        // Generate random tables/relations first so that select items can refer to them.
-        let from = self.gen_from(with_tables);
-        let selection = self.gen_where();
-        let group_by = self.gen_group_by();
-        let having = self.gen_having(!group_by.is_empty());
-        let (select_list, schema) = self.gen_select_list();
-        let select = Select {
-            distinct: Distinct::All,
-            projection: select_list,
-            from,
-            lateral_views: vec![],
-            selection,
-            group_by,
-            having,
-        };
-        (select, schema)
-    }
-
-    fn gen_select_list(&mut self) -> (Vec<SelectItem>, Vec<Column>) {
-        let items_num = self.rng.gen_range(1..=4);
-        let can_agg = self.flip_coin();
-        let context = SqlGeneratorContext::new_with_can_agg(can_agg);
-        (0..items_num)
-            .map(|i| self.gen_select_item(i, context))
-            .unzip()
-    }
-
-    fn gen_select_item(&mut self, i: i32, context: SqlGeneratorContext) -> (SelectItem, Column) {
-        use DataTypeName as T;
-        let ret_type = *[
-            T::Boolean,
-            T::Int16,
-            T::Int32,
-            T::Int64,
-            T::Decimal,
-            T::Float32,
-            T::Float64,
-            T::Varchar,
-            T::Date,
-            T::Timestamp,
-            T::Timestampz,
-            T::Time,
-            T::Interval,
-        ]
-        .choose(&mut self.rng)
-        .unwrap();
-
-        let expr = self.gen_expr(ret_type, context);
-
-        let alias = format!("col_{}", i);
-        (
-            SelectItem::ExprWithAlias {
-                expr,
-                alias: Ident::new(alias.clone()),
-            },
-            Column {
-                name: alias,
-                data_type: ret_type,
-            },
-        )
-    }
-
-    fn gen_from(&mut self, with_tables: Vec<Table>) -> Vec<TableWithJoins> {
-        let mut from = if !with_tables.is_empty() {
-            let with_table = with_tables
-                .choose(&mut self.rng)
-                .expect("with tables should not be empty");
-            vec![create_table_with_joins_from_table(with_table)]
-        } else {
-            let (rel, tables) = self.gen_from_relation();
-            self.add_relations_to_context(tables);
-            vec![rel]
-        };
-
-        if self.is_mview {
-            // TODO: These constraints are workarounds required by mview.
-            // Tracked by: <https://github.com/risingwavelabs/risingwave/issues/4024>.
-            assert!(!self.tables.is_empty());
-            return from;
-        }
-        let mut lateral_contexts = vec![];
-        for _ in 0..self.tables.len() {
-            if self.flip_coin() {
-                let (table_with_join, mut table) = self.gen_from_relation();
-                from.push(table_with_join);
-                lateral_contexts.append(&mut table);
-            }
-        }
-        self.add_relations_to_context(lateral_contexts);
-        from
-    }
-
-    fn gen_where(&mut self) -> Option<Expr> {
-        if self.flip_coin() {
-            let context = SqlGeneratorContext::new_with_can_agg(false);
-            Some(self.gen_expr(DataTypeName::Boolean, context))
-        } else {
-            None
-        }
-    }
-
-    fn gen_group_by(&mut self) -> Vec<Expr> {
-        let mut available = self.bound_columns.clone();
-        if !available.is_empty() {
-            available.shuffle(self.rng);
-            let n_group_by_cols = self.rng.gen_range(1..=available.len());
-            let group_by_cols = available.drain(0..n_group_by_cols).collect_vec();
-            self.bound_columns = group_by_cols.clone();
-            group_by_cols
-                .into_iter()
-                .map(|c| Expr::Identifier(Ident::new(c.name)))
-                .collect_vec()
-        } else {
-            vec![]
-        }
-    }
-
-    fn gen_having(&mut self, have_group_by: bool) -> Option<Expr> {
-        if have_group_by & self.flip_coin() {
-            let context = SqlGeneratorContext::new();
-            Some(self.gen_expr(DataTypeName::Boolean, context))
-        } else {
-            None
-        }
-    }
-
-    /// 50/50 chance to be true/false.
-    fn flip_coin(&mut self) -> bool {
-        self.rng.gen_bool(0.5)
-    }
-
-    /// Provide recursion bounds.
-    pub(crate) fn can_recurse(&mut self) -> bool {
-        self.rng.gen_bool(0.3)
-    }
-}
+pub use crate::sql_gen::{print_function_table, Table};
 
 /// Generate a random SQL string.
 pub fn sql_gen(rng: &mut impl Rng, tables: Vec<Table>) -> String {
@@ -498,6 +57,7 @@ pub fn parse_sql(sql: &str) -> Vec<Statement> {
     Parser::parse_sql(sql).unwrap_or_else(|_| panic!("Failed to parse SQL: {}", sql))
 }
 
+/// Extract relevant info from CREATE TABLE statement, to construct a Table
 pub fn create_table_statement_to_table(statement: &Statement) -> Table {
     match statement {
         Statement::CreateTable { name, columns, .. } => Table {
@@ -506,36 +66,4 @@ pub fn create_table_statement_to_table(statement: &Statement) -> Table {
         },
         _ => panic!("Unexpected statement: {}", statement),
     }
-}
-
-fn is_division_by_zero_err(db_error: &str) -> bool {
-    db_error.contains(&ExprError::DivisionByZero.to_string())
-}
-
-fn is_numeric_out_of_range_err(db_error: &str) -> bool {
-    db_error.contains(&ExprError::NumericOutOfRange.to_string())
-}
-
-/// Skip queries with unimplemented features
-fn is_unimplemented_error(db_error: &str) -> bool {
-    db_error.contains("Feature is not yet implemented")
-}
-
-/// This error occurs because we test `implicit` casts as well,
-/// generated expressions may be ambiguous as a result,
-/// if there are multiple candidates signatures.
-/// Additionally.
-fn not_unique_error(db_error: &str) -> bool {
-    db_error.contains("is not unique\
-    HINT:  Could not choose a best candidate function. You might need to add explicit type casts.")
-}
-
-/// Certain errors are permitted to occur. This is because:
-/// 1. It is more complex to generate queries without these errors.
-/// 2. These errors seldom occur, skipping them won't affect overall effectiveness of sqlsmith.
-pub fn is_permissible_error(db_error: &str) -> bool {
-    is_numeric_out_of_range_err(db_error)
-        || is_division_by_zero_err(db_error)
-        || is_unimplemented_error(db_error)
-        || not_unique_error(db_error)
 }
