+++ src/tests/sqlsmith/src/sql_gen/mod.rs
@@ -1,0 +1,217 @@
+// Copyright 2022 Singularity Data
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#![feature(let_chains)]
+
+use std::vec;
+
+use itertools::Itertools;
+use rand::prelude::SliceRandom;
+use rand::Rng;
+use risingwave_common::types::DataTypeName;
+use risingwave_frontend::bind_data_type;
+use risingwave_sqlparser::ast::{
+    BinaryOperator, ColumnDef, Cte, Distinct, Expr, Ident, Join, JoinConstraint, JoinOperator,
+    ObjectName, OrderByExpr, Query, Select, SelectItem, SetExpr, Statement, TableWithJoins, With,
+};
+use risingwave_sqlparser::parser::Parser;
+
+mod expr;
+pub use expr::print_function_table;
+use risingwave_expr::ExprError;
+
+mod query;
+mod relation;
+mod scalar;
+mod time_window;
+mod utils;
+
+#[derive(Clone, Debug)]
+pub struct Table {
+    pub name: String,
+    pub columns: Vec<Column>,
+}
+
+impl Table {
+    pub fn new(name: String, columns: Vec<Column>) -> Self {
+        Self { name, columns }
+    }
+
+    pub fn get_qualified_columns(&self) -> Vec<Column> {
+        self.columns
+            .iter()
+            .map(|c| Column {
+                name: format!("{}.{}", self.name, c.name),
+                data_type: c.data_type,
+            })
+            .collect()
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct Column {
+    name: String,
+    data_type: DataTypeName,
+}
+
+impl From<ColumnDef> for Column {
+    fn from(c: ColumnDef) -> Self {
+        Self {
+            name: c.name.real_value(),
+            data_type: bind_data_type(&c.data_type).unwrap().into(),
+        }
+    }
+}
+
+#[derive(Copy, Clone)]
+pub(crate) struct SqlGeneratorContext {
+    can_agg: bool, // This is used to disable agg expr totally,
+    // Used in top level, where we want to test queries
+    // without aggregates.
+    inside_agg: bool, // TODO: Maybe we can get rid of this???
+    inside_explicit_cast: bool,
+}
+
+impl SqlGeneratorContext {
+    pub fn new() -> Self {
+        SqlGeneratorContext {
+            can_agg: true,
+            inside_agg: false,
+            inside_explicit_cast: false,
+        }
+    }
+
+    pub fn new_with_can_agg(can_agg: bool) -> Self {
+        Self {
+            can_agg,
+            inside_agg: false,
+            inside_explicit_cast: false,
+        }
+    }
+
+    pub fn set_inside_agg(self) -> Self {
+        Self {
+            inside_agg: true,
+            ..self
+        }
+    }
+
+    pub fn set_inside_explicit_cast(self) -> Self {
+        Self {
+            inside_explicit_cast: true,
+            ..self
+        }
+    }
+
+    pub fn can_implicit_cast(self) -> bool {
+        !self.inside_explicit_cast
+    }
+
+    pub fn can_gen_agg(self) -> bool {
+        self.can_agg && !self.inside_agg
+    }
+}
+
+pub(crate) struct SqlGenerator<'a, R: Rng> {
+    tables: Vec<Table>,
+    rng: &'a mut R,
+
+    /// Relation ID used to generate table names and aliases
+    relation_id: u32,
+
+    /// is_distinct_allowed - Distinct and Orderby/Approx.. cannot be generated together among agg
+    ///                       having and
+    /// When this variable is true, it means distinct only
+    /// When this variable is false, it means orderby and approx only.
+    is_distinct_allowed: bool,
+
+    /// Relations bound in generated query.
+    /// We might not read from all tables.
+    bound_relations: Vec<Table>,
+
+    /// Columns bound in generated query.
+    /// May not contain all columns from Self::bound_relations.
+    /// e.g. GROUP BY clause will constrain bound_columns.
+    bound_columns: Vec<Column>,
+
+    /// SqlGenerator can be used in two execution modes:
+    /// 1. Generating Query Statements.
+    /// 2. Generating queries for CREATE MATERIALIZED VIEW.
+    ///    Under this mode certain restrictions and workarounds are applied
+    ///    for unsupported stream executors.
+    is_mview: bool,
+}
+
+/// Generators
+impl<'a, R: Rng> SqlGenerator<'a, R> {
+    pub(crate) fn new(rng: &'a mut R, tables: Vec<Table>) -> Self {
+        let is_distinct_allowed = rng.gen_bool(0.5);
+        SqlGenerator {
+            tables,
+            rng,
+            relation_id: 0,
+            is_distinct_allowed,
+            bound_relations: vec![],
+            bound_columns: vec![],
+            is_mview: false,
+        }
+    }
+
+    pub(crate) fn new_for_mview(rng: &'a mut R, tables: Vec<Table>) -> Self {
+        // distinct aggregate is not allowed for MV
+        SqlGenerator {
+            tables,
+            rng,
+            relation_id: 0,
+            is_distinct_allowed: false,
+            bound_relations: vec![],
+            bound_columns: vec![],
+            is_mview: true,
+        }
+    }
+
+    pub(crate) fn gen_stmt(&mut self) -> Statement {
+        let (query, _) = self.gen_query();
+        Statement::Query(Box::new(query))
+    }
+
+    pub(crate) fn gen_mview(&mut self, name: &str) -> (Statement, Table) {
+        let (query, schema) = self.gen_query();
+        let query = Box::new(query);
+        let table = Table {
+            name: name.to_string(),
+            columns: schema,
+        };
+        let name = ObjectName(vec![Ident::new(name)]);
+        let mview = Statement::CreateView {
+            or_replace: false,
+            materialized: true,
+            name,
+            columns: vec![],
+            query,
+            with_options: vec![],
+        };
+        (mview, table)
+    }
+
+    /// 50/50 chance to be true/false.
+    fn flip_coin(&mut self) -> bool {
+        self.rng.gen_bool(0.5)
+    }
+
+    /// Provide recursion bounds.
+    pub(crate) fn can_recurse(&mut self) -> bool {
+        self.rng.gen_bool(0.3)
+    }
+}
